fix splash always showing issue
