+++ src/screens/Splash/index.tsx
@@ -1,16 +1,10 @@
-import React, {FC, useEffect, useMemo, useRef} from 'react';
+import React, {FC, useEffect, useRef} from 'react';
 import {Animated, Easing} from 'react-native';
 
-import {useNavigation} from '@react-navigation/native';
 import styled from 'styled-components/native';
 
 import {SPLASH_IMAGE} from '~/assets';
 import BackgroundCommon from '~/components/BackgroundCommon';
-import {
-  ApplicationStackName,
-  ApplicationStackNavigationProps,
-} from '~/navigation/Application';
-import {useTypedSelector} from '~/store';
 import {ContainerStyle} from '~/utils/Styles';
 
 const Container = styled.View`
@@ -22,38 +16,7 @@ const SplashImage = styled(Animated.Image)<{opacity: Animated.Value}>`
 `;
 
 const Splash: FC = () => {
-  const {replace} = useNavigation<ApplicationStackNavigationProps>();
-  const {
-    authData: {isLoggedIn},
-    userData: {userName, categoryAndTags, stardustRecords},
-  } = useTypedSelector(({user}) => user);
-  const hasUserData = useMemo(
-    () =>
-      !!userName && categoryAndTags.length > 0 && stardustRecords.length > 0,
-    [userName, categoryAndTags, stardustRecords],
-  );
-
   const opacity = useRef(new Animated.Value(0)).current; // Initial value for opacity: 0
-
-  useEffect(() => {
-    const timeoutId: NodeJS.Timeout = setTimeout(() => {
-      if (!isLoggedIn) {
-        return replace(ApplicationStackName.AUTH);
-      }
-
-      if (!hasUserData) {
-        return replace(ApplicationStackName.MAIN);
-      }
-
-      // TODO: add redirection to main page
-    }, 3000);
-
-    return () => {
-      if (timeoutId) {
-        clearTimeout(timeoutId);
-      }
-    };
-  }, []);
 
   useEffect(() => {
     Animated.timing(opacity, {
